# internal/driver/glfw: correct handle goroutineID