--- a/internal/driver/glfw/driver_goxjs.go
+++ b/internal/driver/glfw/driver_goxjs.go
@@ -3,6 +3,6 @@
 
 package glfw
 
-func goroutineID() int {
+func goroutineID() uint64 {
 	return mainGoroutineID
 }