### Description:

The current goroutine ID is handled with two issues:

1. main goroutine is assumed to be 1, but it may not be true;
2. goroutine ID is typed as int rather than uin64; this may cause overflow.

Let's handle them gracefully.

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [x] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.