@Jacalz 2022-05-01
Wonderful. I’ll have a look at this tomorrow and also see about benchmarking it because I’m curious :)