@changkun 2022-05-01
Benchmark is not much interesting, though. The implementation is bounded by the `runtime.Stack` performance. Optimizing conversions are just a very few ns faster. I think it would be better for us to fix the current misuse and potential danger in the implementation:

```
name                      old time/op  new time/op  delta
GoroutineID/sequential-8  1.88µs ± 0%  1.78µs ± 0%  -5.38%  (p=0.000 n=4+5)
GoroutineID/concurrent-8  2.66µs ± 2%  2.53µs ± 2%  -5.07%  (p=0.008 n=5+5)
```

benchmark: https://go.dev/play/p/ZuN6iZMj2RT