@Jacalz 2022-05-02
> Optimizing conversions are just a very few ns faster. I think it would be better for us to fix the current misuse and potential danger in the implementation:

Absolutely. I just happen to be curious :)